Bump minor version number for new 1.1 release
This release adds support for platforms that prefix symbols with an
extra leading underscore, or for mixtures of prefixed and non-prefixed
symbols.  Both `_Z5printi` and `__Z5printi` demangle identically to
`print(int)`, regardless of the host platform’s prefixing conventions.

Special thanks to GitHub user @kastiglione for reporting this problem
(issue #2) and helping me identify a robust solution.