Don't fontify matches that demangle to themselves
This can happen when a run of text matches our regular expression but is
not actually a valid mangled symbol, such as "_Z1fi2".  The mangled
symbol grammar is complex enough that hitting it perfectly with the
regular expression will be difficult or impossible.  Instead, we now
accept the possibility that regexp-matched text may not demangle.  That
comes back from c++filt looking like text that demangled to itself.  If
that's what we find, then leave the matched text alone, since there's
really nothing to see.