Try harder to preserve the match-data we need
When running Travis CI tests under Emacs 24.5, the match-data contains
integers instead of markers.  That suggests that we are picking up match
data from some sort of string-matching operation, rather than a
buffer-matching operation.  I suspect that this may have something to do
with the special fontify-and-drain-demangler font lock function we are
using during testing.  Hopefully we are now being more robust about not
letting other searches pollute the match-data we need here.