Don't make `demangle-show-as` buffer-local
If a user wants this to be buffer-local, then he or she can use
`make-variable-buffer-local` as needed.  And any buffer-local variables
set in a "Local variables:" section or in a ".dir-locals.el" file will
already be buffer local.  Barring any of that, though, we assume that
`demangle-show-as` has a single global setting.  Changing it anywhere
changes it in all `demangle-mode` buffers, which will be re-fontified
accordingly.

The Emacs Lisp manual explicitly recommends this approach:

    *Warning:* Don’t assume that you should use
    ‘make-variable-buffer-local’ for user-option variables, simply
    because users _might_ want to customize them differently in
    different buffers.  Users can make any variable local, when they
    wish to.  It is better to leave the choice to them.

I'm convinced.