Revert font-lock-extra-managed-props when disabling mode
Internally, this works by allowing ourselves to add duplicate values to
font-lock-extra-managed-props when turning demangle-mode on.  When
turning the mode off later, we restrict ourselves to removing only one
copy of each of the properties we added earlier.  For example, if
help-echo was already on this list, then we will add a second copy when
turning demangle-mode on, and we'll remove just one of the two copies
when turning demangle-mode off.  Neat!