Better interactive prompting for demangle-show-as
Previously, entering anything other than "demangled" or "mangled" led to
an unfriendly error message.  Now, these are the only acceptable
responses.  However, we also have better completion, so that an input as
short as "d" or "m" expands to "demangled" or "mangled", respectively.
An input like "f", however, will report "[No match]" and continue
prompting for a selection.  Pressing <return> with no text entered at
all defaults to toggling the show-as style.