Adapt to change in faceup-test-font-lock-buffer API
Commit 688b487ad0a78c8707c5aded50e1d85551270034 of faceup changed
faceup-test-font-lock-buffer to let the MODE argument be a list of
mode-setting functions instead of a single mode-setting function.
If MODE is a list, then faceup-test-font-lock-buffer calls each
function in that list.

Unfortunately, when we call faceup-test-font-lock-buffer, we are often
passing a list of the form "(closure ...)", because this is the
representation of a lambda when using lexical scoping.  This causes
faceup-test-font-lock-buffer to try to call each item in this list,
starting with "(closure)", which is not what we want to have happen.

The work-around is to always call faceup-test-font-lock-buffer with a
list of our own creation, even if it's a list of just one item.  Thus,
"((closure ...))" instead of "(closure ...)".